libs/br-variable-api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "br-variable-api"
version = "0.1.0"
description = "Variable API for B+R documentation - access variables with URL queries"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Softreck", email = "info@softreck.dev"}
]
dependencies = [
    "br-core>=0.1.0",
    "br-data-sources>=0.1.0",
    "fastapi>=0.104",
    "pydantic>=2.0",
    "structlog>=23.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]

[tool.setuptools.packages.find]
where = ["src"]
